refactor: Refactor packaging of all entities
