+++ src/main/java/com/yum/yumyums/entity/seller/Discount.java
@@ -1,4 +1,4 @@
-package com.yum.yumyums.entity;
+package com.yum.yumyums.entity.seller;
 
 import jakarta.persistence.*;
 import lombok.Getter;
