+++ src/main/java/com/yum/yumyums/entity/user/PointList.java
@@ -1,5 +1,6 @@
-package com.yum.yumyums.entity;
+package com.yum.yumyums.entity.user;
 
+import com.yum.yumyums.entity.user.Member;
 import jakarta.persistence.*;
 import lombok.Getter;
