+++ src/main/java/com/yum/yumyums/entity/user/SearchFilter.java
@@ -1,4 +1,4 @@
-package com.yum.yumyums.entity;
+package com.yum.yumyums.entity.user;
 
 import com.yum.yumyums.enums.FilterCategory;
 import jakarta.persistence.*;
